+++ packages/ui/src/components/internal/icons/Add.svelte
